requirements.txt
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
async-timeout==3.0.1
asyncpg==0.22.0
asyncssh==2.5.0
attrs==20.3.0
backoff==1.10.0
base58==2.1.0
black==21.8b0
Cerberus==1.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chevron==0.14.0
click==7.1.2
click-shell==2.0
cryptography==3.4.6
Cython==0.29.23
diskcache==5.2.1
ecdsa==0.16.1
fuuid==0.1.0
gevent==21.1.2
gnureadline==8.0.0
greenlet==1.0.0
h11==0.12.0
idna==2.10
itsdangerous==1.1.0
jsonschema==2.6.0
jwcrypto==0.8
lru-dict==1.1.7
montydb==2.3.12
multidict==5.1.0
mypy-extensions==0.4.3
orjson==3.5.1
pathspec==0.9.0
platformdirs==2.3.0
prompt-toolkit==1.0.14
psutil==5.8.0
py-cpuinfo==8.0.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pycurl==7.43.0.6
Pygments==2.8.1
PyInquirer==1.0.3
PyJWT==2.1.0
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-geoip==1.2
python-geoip-geolite2==2015.303
python-geoip-python3==1.3
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.15.0
sshpubkeys==3.3.1
supervisor==4.2.2
terminaltables==3.1.0
texttable==1.6.3
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.4
uvicorn==0.13.4
wcwidth==0.2.5
websocket==0.2.1
websockets==8.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.3.0
